// frontend/src/api/model/documentTemplateFormFieldType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * Field type
 */
export type DocumentTemplateFormFieldType = typeof DocumentTemplateFormFieldType[keyof typeof DocumentTemplateFormFieldType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DocumentTemplateFormFieldType = {
  TEXT: 'TEXT',
  MULTILINE_TEXT: 'MULTILINE_TEXT',
  CHECKBOX: 'CHECKBOX',
  RADIO: 'RADIO',
  SELECT: 'SELECT',
  UNKNOWN: 'UNKNOWN',
} as const;