// frontend/src/api/model/documentTemplateFormField.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DocumentTemplateFormFieldType } from './documentTemplateFormFieldType';

/**
 * Single parsed field inside a document template form map
 */
export interface DocumentTemplateFormField {
  /** Field key */
  key?: string;
  /** Field type */
  type?: DocumentTemplateFormFieldType;
  /** Possible values for selectable controls */
  possibleValues?: string[];
}